Beginnings of vfs_initrd driver

+++ src/kernel/multiboot.c
@@ -18,7 +18,7 @@ struct multiboot make_multiboot_physical(struct multiboot *old)
 	kassert((size_t)mb.mods_addr > 0xc0000000, "mb.mods_addr PHYSICAL");
 	for (int i = 0; i < mb.mods_count + 1; i++)
 	{
-		TO_VIRT(mb.mods_addr[i], uint *);
+		TO_VIRT(mb.mods_addr[i], uchar *);
 	}
 
 	return mb;
